authorization

+++ backend/controllers/todoController.js
@@ -1,10 +1,12 @@
 const asyncHandler = require("express-async-handler");
 const Todo = require("../models/todoModel");
+const User = require("../models/userModel");
+
 //@desc Get Todos
 //@route GET/api/todos
 //@access Private
 const getTodos = asyncHandler(async (req, res) => {
-  const todos = await Todo.find();
+  const todos = await Todo.find({ user: req.user.id });
   res.status(200).json(todos);
 });
 
@@ -19,6 +21,7 @@ const setTodo = asyncHandler(async (req, res) => {
 
   const todo = await Todo.create({
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.status(200).json(todo);
@@ -33,6 +36,21 @@ const updateTodo = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Todo not found");
   }
+
+  const user = await User.findById(req.user.id);
+
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //make sure the logged is user matches the todo user
+  if (todo.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   const updatedTodo = await Todo.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -50,26 +68,22 @@ const deleteTodo = asyncHandler(async (req, res) => {
     throw new Error("Todo not found");
   }
 
+  const user = await User.findById(req.user.id);
+
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //make sure the logged is user matches the todo user
+  if (todo.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   await todo.deleteOne();
   res.status(200).json({ id: req.params.id });
-
-  //   const todo = await Todo.findById(req.params.id);
-  //   if (!todo) {
-  //     res.status(400);
-  //     throw new Error("Todo not found");
-  //   }
-  //   //check for user
-  //   if (!req.user) {
-  //     res.status(401);
-  //     throw new Error("User not found");
-  //   }
-  //   //make sure the logged in user matches the togo user
-  //   if (todo.user.toString() !== req.user.id) {
-  //     res.status(401);
-  //     throw new Error("User not authorized");
-  //   }
-  //   await todo.deleteOne();
-  //   res.status(200).json({ id: req.params.id });
 });
 
 module.exports = {
